requirements.txt
Pillow==9.5.0
pytorch_lightning==2.0.0
PyYAML==6.0
torch==2.0.0
torchvision==0.15.0
